requirements.txt
fastapi
uvicorn
python-multipart
python-docx
reportlab
openai
python-dotenv
pdfplumber
